Add OneBit mode to HARDWARE menu.  OneBit mode can help if your radio has a slow inverter circuit by changing the UART sampling behaviour.
* It was found that the X9D+ and X7 (maybe others) have a slow inveter
that messes up reception of fast (400,000baud) UART signals.  On the
scope the rise time of the signal is very long.
* See RM0033 24.3.3 Receiver / Selecting the proper oversampling method.
Page 631.


Also note that reducing the pull-up resistor from 10k to 1k at R82
(X9D+) / R406 (X7) will also help with the rise times.  However on the
X7 changing the resistor is not enough, OneBit is needed too.