[nv14] Make Outputs widget work with two column layouts (#1454)
Changed code to allow one column display of outputs already with a width of 100. This fixes the initally reported problem

Reason: On the Nirvana the screen width is 320 pixels. Thus the widget width is less than 150 pixels, if you select a two column out layout.
The rendering of the outputs looks good with a width of 100px, so there is no need for the "150" - we can start displaying from 100. Should make no difference to the radios with width 480px.