Improved gt911 code
* fixes axis reversal config for gt911
* if version code is different write new config

The datasheet of the gt911 is misleading: a new config is stored
in the chip if the version code is higher OR if the config
is different. Here: the config for TX16S and X12S are different (axis reversal)
so the gt911 would accept the new even with same version code.

This driver originally checks if new version code
is higher than stored version code. This would it make impossible
to move a touch-dsiplay back from X12S to TX16S as then the new config would
not be written