fix: race condition on mixer task handle
PR #3451 introduced a change that caused the mixer task to be started out of the UI task, which introduces a race condition `mixerTaskId.rtos_handle`, as the mixer scheduler might trigger before `xTaskCreateStatic` returns and the task handle is assigned.

This PR fixes the issue by delaying the start of the mixer scheduler until `xTaskCreateStatic` returned and the task handler has been properly assigned.