Fix Flysky Hall Gimbals not working
This fixes https://github.com/EdgeTX/edgetx/issues/4859

On Horus Radios with PWM_STICKS the optional support for the Flysky Hall gimbals is disfunctional since 2.10.
The problem is that both use the same GPIOs. With sticksPwmDetect() the GPIOs get blocked thus flysky_gimbal_init() does not work any more. So I turned the order to first look for the Flysky Gimbals and go to PWM_Sticks if not found.